+++ CHANGELOG.md
@@ -1,1 +1,4 @@
 # v2.11.32
+
+* mappings that don't have the `items` method will be 
+  correctly processed by `jsonify`